+++ lib/hcp/account.rb
@@ -16,6 +16,9 @@ module Hcp
     # @return [Visits] visits of the location, read off its jobs.
     def visits = Visits.new client: @client
 
+    # @return [Technicians] employees of the location, walked a page at a time.
+    def technicians = Technicians.new client: @client
+
     # @return [Leads] leads of the location, to open and to move.
     def leads = Leads.new client: @client
   end